% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_missing_days}
\alias{add_missing_days}
\title{Add Missing Days to Data}
\usage{
add_missing_days(data)
}
\arguments{
\item{data}{A dataframe containing \code{id}, \code{date}, and \code{load} columns.}
}
\value{
A dataframe with missing days added.
}
\description{
Ensures the data has complete daily entries for each ID.
}
