% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Features}
\alias{Features}
\title{Feature Class}
\description{
R6 class representing a feature with raw and processed data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Feature-new}{\code{Features$new()}}
\item \href{#method-Feature-get_data}{\code{Features$get_data()}}
\item \href{#method-Feature-clone}{\code{Features$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-new"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Features$new(data_raw = data.frame(), data_processed = data.frame())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Features$get_data(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Features$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
