% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Load}
\alias{Load}
\title{Load Class}
\description{
Handles load data processing, including scaling and NA handling.
}
\section{Super class}{
\code{MILS::Feature} -> \code{Load}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Load-new}{\code{Load$new()}}
\item \href{#method-Load-process}{\code{Load$process()}}
\item \href{#method-Load-clone}{\code{Load$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="MILS" data-topic="Feature" data-id="get_data"><a href='../../MILS/html/Feature.html#method-Feature-get_data'><code>MILS::Feature$get_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Load-new"></a>}}
\if{latex}{\out{\hypertarget{method-Load-new}{}}}
\subsection{Method \code{new()}}{
Initializes a Load object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Load$new(data_raw = data.frame(), data_processed = data.frame())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_raw}}{A dataframe with raw data.}

\item{\code{data_processed}}{A dataframe with processed data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Load-process"></a>}}
\if{latex}{\out{\hypertarget{method-Load-process}{}}}
\subsection{Method \code{process()}}{
Processes the raw load data, handles missing values, and scales the load.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Load$process()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Load-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Load-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Load$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
