% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_print}
\alias{my_print}
\title{Print Current Time with Message}
\usage{
my_print(string)
}
\arguments{
\item{string}{A message string to print.}
}
\value{
None
}
\description{
Print Current Time with Message
}
