% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pipeline}
\alias{pipeline}
\title{Data Processing Pipeline}
\usage{
pipeline(features)
}
\arguments{
\item{features}{A list of Feature objects with processed data.}
}
\value{
A combined dataframe with processed features.
}
\description{
Processes a list of features and combines them into a dataframe.
}
