% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_emiss}
\alias{plot_emiss}
\title{Plot Emission Matrix}
\usage{
plot_emiss(emiss)
}
\arguments{
\item{emiss}{A dataframe representing the emission matrix.}
}
\value{
A ggplot object.
}
\description{
Plots an emission matrix with means and standard deviations.
}
