% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_trans}
\alias{plot_trans}
\title{Plot Transition Graph}
\usage{
plot_trans(links, nodes, emiss)
}
\arguments{
\item{links}{A dataframe of links (edges).}

\item{nodes}{A dataframe of nodes (vertices).}

\item{emiss}{A dataframe representing the emission matrix.}
}
\value{
A network graph plot.
}
\description{
Creates a network graph from transition data.
}
